refactor: 관련된 타입별로 파일 분리
